DREAMWEB: Add titles.cpp, move some methods to titles.cpp and talk.cpp
